refactor: Move GeolocationTrackingProviders in providers folder
no code change
